+++ players/typescript/src/core/operations/DrawRect.ts
@@ -19,7 +19,7 @@ export class DrawRect extends DrawBase4 {
 
     static read(buffer: WireBuffer, operations: Operation[]): void {
         operations.push(new DrawRect(
-            buffer.readFloat(), buffer.readFloat(), buffer.readFloat(), buffer.readFloat()
+            buffer.readInt(), buffer.readInt(), buffer.readInt(), buffer.readInt()
         ));
     }
 }
